alternator: add binding
Alternator is ScyllaDB's DynamoDB compatible API,
since it is cluster based, the Dynamodb client is a little
different than the classic DynamoDB client, it does load
balancing on the client side.
To this end, we replicated the dynamodb-binding infrastructure
and replaced only the client creation code.
We could also do it differently and add a parameter that will
create the client differently but we wanted to leave the original
dynamodb code untouched to avoid confusion.