dynamodb: use default AWS credentials, not a "properties file"
Currently, the DynamoDB front-end requires a dynamodb.awsCredentialsFile
parameter to specify the credentials to use to make requests to DynamoDB.
This uses AWS library PropertiesCredentials() feature, which is limited,
and for example doesn't support STS token strings.

We should simply remove this PropertiesCredentials code, and let the
AWS library use its default configuration mechanism - using a file in
~/.aws/credentials or environment variables - as this is the standard
and full-featured mechanism.

After this patch, the dynamodb.awsCredentialsFile option is no longer needed
or supported. Note that the format of ~/.aws/credentials, which you now
need to create, is slightly different.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>