ycsb.core: Support java version later than 8
On Java 9 and later, some of the xml related APIs were removed
from the default JAVA library and it makes some of the xml.bind
calls not to work since they don't exist anymore.
Here we add a package that provides them so ycsb can work also on
those later version.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>