ScrambledZipfianGenerator: Add a main method for generating test sets
Instead of assuming that we know how is the ScrambledZipfianGenerator
class works, we simply add a main class that can be run from command
line and generate a requested amount of samples from a requested range.
Than it can be used to analyze after the fact how those sets looks like
and how are the frequencies really distributes.

NOTE: This is a static main method that does not affect YCSB at all.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>