RawMeasurementOne: dump to the output file every status interval
It has been observed during long runs with high throughput that using
raw measurements becomes impractical due to it using too much heap space
that eventually causes YCSB to crash.
Each sample takes at least 16 bytes of memory which translates to ~1.5MB
of RAM for each 100Kops. If you run a highly performant DB.
It is not unreasonable that after several tens of minutes the RAM
consumption will be multiple GBs.
This commit dumps the accumulated measurements during the period between
status prints. It keeps the RAM consumption constant (on steady state)
and also has the advantage of having almost all the results of a long
run if it crashes in the middle.
The invariants are:
Every dump starts with a header. Between two headers there is only one
measurement types (they are not intermixed).

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>